refactor deck and upload
